+++ 2024-Competition-Code/src/main/java/frc/robot/commands/TargetAprilTag.java
@@ -4,7 +4,7 @@ import edu.wpi.first.math.geometry.Translation2d;
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj.XboxController;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.PhotonVision.*;
+import frc.robot.subsystems.photonvision.*;
 import frc.robot.subsystems.swerve.SwerveDrive;
